backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
sse-starlette>=1.6.1
pydantic>=2.0.0
python-multipart>=0.0.6
